branches

+++ src/main/java/objects/AccountHistory.java
@@ -8,7 +8,7 @@ public class AccountHistory {
     private String operationType;
     private float sumChanging;
 
-    public AccountHistory(long balance, LocalDateTime time, String operationType, long sumChanging) {
+    public AccountHistory(float balance, LocalDateTime time, String operationType, float sumChanging) {
         this.balance = balance;
         this.time = time;
         this.operationType = operationType;
